+++ frontend/public/extend/devconsole/components/ImportFlowForm/__tests__/ImportFlowForm.spec.tsx
@@ -4,9 +4,7 @@ import { ImportFlowForm } from './../ImportFlowForm';
 
 describe('ImportFlowForm: ', () => {
   const props = {
-    namespace: {
-      data: [],
-    },
+    activeNamespace: 'default',
   };
 
   it('renders', () => {
